rearranged the new code
